requirements.txt
requests>=2.31.0
tqdm>=4.66.1
langchain-community>=0.0.27
langchain>=0.1.9
langgraph>=0.0.11
python-dotenv>=1.0.0 